# assets/openapi/sonos-hub.v1.yaml
openapi: 3.0.3
info:
  title: Sonos Smart Routines Hub API
  version: '1.0.0'
  description: |
    REST API for the Sonos Smart Routines Hub - a LAN-only home automation system for
    Sonos speakers. Provides device discovery, scene/routine management, music playback
    control, and integration with Apple Music. Designed for local-first operation with
    observability via Prometheus metrics and structured JSON logging.
servers:
  - url: http://hub.local:8787
...
paths:
...
  /v1/auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      description: Exchange a refresh token for a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthAccessTokenResponse' }


  # =========================================================================
  # DASHBOARD ENDPOINTS
  # =========================================================================
...
  /v1/sonos-cloud/auth/callback:
    get:
      operationId: handleSonosCloudCallback
      tags: [sonos-cloud]
      summary: OAuth callback handler
      description: Handles OAuth redirect from Sonos, exchanges authorization code for tokens
      parameters:
        - in: query
          name: code
          description: Authorization code from Sonos
          schema:
            type: string
        - in: query
          name: state
          schema:
            type: string
          description: CSRF state parameter
        - in: query
          name: error
          schema:
            type: string
          description: Error code if authorization failed
        - in: query
          name: error_description
          schema:
            type: string
          description: Human-readable error description
      responses:
        '200':
          description: Success HTML page
          content:
            text/html:
              schema:
                type: string
        '400':
          description: OAuth error or invalid state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SonosCloudErrorResponse'
        '500':
          description: Token exchange failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SonosCloudErrorResponse'
...
  /v1/sonos-cloud/auth/refresh:
    post:
      operationId: refreshSonosCloudToken
      tags: [sonos-cloud]
      summary: Force refresh access token
      description: Manually refresh the Sonos Cloud access token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SonosCloudAuthRefreshResponse'
        '500':
          description: Refresh failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SonosCloudErrorResponse'
  /v1/sonos-cloud/auth/start:
    get:
      operationId: startSonosCloudAuth
      tags: [sonos-cloud]
      summary: Get OAuth authorization URL
      description: Returns the Sonos OAuth authorization URL and state parameter for CSRF protection
      parameters:
        - in: query
          name: redirect
          description: If 'true', redirects browser directly to Sonos OAuth
          schema:
            type: string
            enum: ['true', 'false']
      responses:
        '200':
          description: Returns authorization URL and state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SonosCloudAuthStartResponse'
        '302':
          description: Redirects to Sonos OAuth (when redirect=true)
...
  /v1/sonos-cloud/groups:
    get:
      operationId: listSonosCloudGroups
      tags: [sonos-cloud]
      summary: List groups in household
      description: Returns all speaker groups in the specified household
      parameters:
        - in: query
          name: householdId
          description: Household ID (uses cached household if omitted)
          schema:
            type: string
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SonosCloudGroupsResponse'
        '400':
          description: Missing household ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SonosCloudErrorResponse'
        '500':
          description: API error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SonosCloudErrorResponse'
  /v1/sonos-cloud/households:
    get:
      operationId: listSonosCloudHouseholds
      tags: [sonos-cloud]
      summary: List connected households
      description: Returns all Sonos households associated with the authenticated account
      responses:
        '200':
          description: List of households
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SonosCloudHouseholdsResponse'
        '500':
          description: API error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SonosCloudErrorResponse'
  /v1/sonos-cloud/players:
    get:
      operationId: listSonosCloudPlayers
      tags: [sonos-cloud]
      summary: List players in household
      description: Returns all Sonos players in the specified household
      parameters:
        - in: query
          name: householdId
          description: Household ID (uses cached household if omitted)
          schema:
            type: string
      responses:
        '200':
          description: List of players
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SonosCloudPlayersResponse'
        '400':
          description: Missing household ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SonosCloudErrorResponse'
        '500':
          description: API error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SonosCloudErrorResponse'
  /v1/sonos-cloud/players/{playerId}/audioClip:
    post:
      operationId: playSonosCloudAudioClip
      tags: [sonos-cloud]
      summary: Play audio clip on player
      description: Overlay an audio notification on the player without interrupting current playback
      parameters:
        - in: path
          name: playerId
          description: Sonos player ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SonosCloudAudioClipRequest'
      responses:
        '200':
          description: Audio clip scheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SonosCloudAudioClipResponse'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SonosCloudErrorResponse'
        '500':
          description: API error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SonosCloudErrorResponse'

  # =========================================================================
  # SYSTEM ENDPOINTS
  # =========================================================================
...
components:
  schemas:
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
...
    AuthAccessTokenResponse:
      type: object
      required: [request_id, access_token, expires_in_sec]
      properties:
        request_id: { type: string }
        access_token: { type: string }
        expires_in_sec: { type: integer }
...
    SonosCloudAuthStartResponse:
      type: object
      required: [authUrl, state]
      properties:
        authUrl:
          type: string
          format: uri
          description: Full OAuth authorization URL
        state:
          type: string
          description: CSRF protection state token
...
    SonosCloudAuthRefreshResponse:
      type: object
      required: [success, expiresAt, scope]
      properties:
        success:
          type: boolean
        expiresAt:
          type: string
          format: date-time
        scope:
          type: string
...
    SonosCloudHousehold:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
    SonosCloudHouseholdsResponse:
      type: object
      required: [households]
      properties:
        households:
          type: array
          items:
            $ref: '#/components/schemas/SonosCloudHousehold'
    SonosCloudGroup:
      type: object
      required: [id, name, coordinatorId, playerIds]
      properties:
        id:
          type: string
        name:
          type: string
        coordinatorId:
          type: string
        playerIds:
          type: array
          items:
            type: string
    SonosCloudGroupsResponse:
      type: object
      required: [householdId, groups]
      properties:
        householdId:
          type: string
        groups:
          type: array
          items:
            $ref: '#/components/schemas/SonosCloudGroup'
    SonosCloudPlayer:
      type: object
      required: [id, name, softwareVersion, deviceIds, capabilities]
      properties:
        id:
          type: string
        name:
          type: string
        softwareVersion:
          type: string
        deviceIds:
          type: array
          items:
            type: string
        capabilities:
          type: array
          items:
            type: string
    SonosCloudPlayersResponse:
      type: object
      required: [householdId, players]
      properties:
        householdId:
          type: string
        players:
          type: array
          items:
            $ref: '#/components/schemas/SonosCloudPlayer'
    SonosCloudAudioClipRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Clip name for identification
        streamUrl:
          type: string
          format: uri
          description: URL to audio file (mp3/wav) - must be publicly accessible
        volume:
          type: integer
          minimum: 0
          maximum: 100
          description: Volume level for the clip (0-100)
        priority:
          type: string
          enum: [HIGH, LOW]
          description: HIGH interrupts current audio, LOW can be interrupted
    SonosCloudAudioClipResponse:
      type: object
      required: [success, clipId, name]
      properties:
        success:
          type: boolean
        clipId:
          type: string
          description: Sonos-assigned clip ID
        name:
          type: string
        status:
          type: string
          description: Clip status from Sonos
    SonosCloudErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message

    # =========================================================================
    # System Schemas
    # =========================================================================
...